move inlinePanel function to src/includes and add basic test
- basic unit test is mostly a smoke test that the init functions get called
- add JSDOC to inlinePanel function body
- name some internal functions
